Organize MCP server sources
